+++ examples/get_started/main.go
@@ -49,6 +49,7 @@ func mainView(w *gui.Window) gui.View {
 				Text: "Click to increment counter",
 				Content: []gui.View{
 					gui.Button(gui.ButtonCfg{
+						ID:        "gs_counter",
 						Focusable: true,
 						Shadow: &gui.BoxShadow{
 							Color:      gui.RGBA(0, 0, 64, 64),